// types/cowboy.ts
export interface Cowboy {
  id: number;
  name: string | null;
  new_man_year: number | null;
  city: string | null;
  state: string | null;
  industry: string | null;
  industry_source: string | null;
  org: string | null;
  major: string | null;
  grad_school: string | null;
  high_school: string | null;
  company: string | null;
  role: string | null;
  linkedin: string | null;
  deceased: boolean | null;
  obit: string | null;
}